dom/broadcastchannel/tests/test_broadcastchannel_self.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for BroadcastChannel</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>

<div id="content"></div>

<script type="application/javascript">

function runTest() {
  x = new BroadcastChannel('foo');
  y = new BroadcastChannel('foo');

  function func(e) {
    is(e.target, y, "The target is !x");

    SimpleTest.executeSoon(function() {
      SimpleTest.finish();
    });
  }

  x.onmessage = func;
  y.onmessage = func;

  x.postMessage('foo');
}

SimpleTest.waitForExplicitFinish();
runTest();

</script>
</body>
</html>
